working local eslint rule
