+++ .eslintrc
@@ -1,6 +1,6 @@
 {
   "parser": "@typescript-eslint/parser",
-  "plugins": ["@typescript-eslint", "import"],
+  "plugins": ["@typescript-eslint", "import", "local-rules"],
   "env": {
     "browser": true,
     "node": true,
@@ -52,15 +52,20 @@
             "ignorePackages": true,
             "checkTypeImports": true
           }
-        ]
+        ],
+        "local-rules/require-using-disposable": "error"
       }
     },
     {
-      "files": ["**/__tests__/**/*.[jt]sx", "**/?(*.)+(test).[jt]sx"],
+      "files": ["**/__tests__/**/*.[jt]s", "**/__tests__/**/*.[jt]sx", "**/?(*.)+(test).[jt]s", "**/?(*.)+(test).[jt]sx"],
       "extends": ["plugin:testing-library/react"],
+      "parserOptions": {
+        "project": "./tsconfig.tests.json"
+      },
       "rules": {
         "testing-library/prefer-user-event": "error",
-        "testing-library/no-wait-for-multiple-assertions": "off"
+        "testing-library/no-wait-for-multiple-assertions": "off",
+        "local-rules/require-using-disposable": "error"
       }
     }
   ],
