--- config/eslint-rules/package.json
@@ -1,6 +1,0 @@
-{
-  "type": "module",
-  "scripts": {
-    "test": "node *.test.mjs"
-  }
-}
