+++ eslint-local-rules/fixtures/tsconfig.json
