+++ eslint-local-rules/package.json
@@ -1,0 +1,5 @@
+{
+  "scripts": {
+    "test": "ts-node *.test.ts"
+  }
+}
