+++ eslint-local-rules/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "extends": "@tsconfig/node20/tsconfig.json",
   "compilerOptions": {
-    "allowJs": true
+    "allowJs": true,
+    "noEmit": true,
   }
 }
