+++ package.json
@@ -128,7 +128,7 @@
     "@types/use-sync-external-store": "0.0.3",
     "@typescript-eslint/eslint-plugin": "6.6.0",
     "@typescript-eslint/parser": "6.6.0",
-    "@typescript-eslint/rule-tester": "^6.6.0",
+    "@typescript-eslint/rule-tester": "6.6.0",
     "@typescript-eslint/types": "6.6.0",
     "@typescript-eslint/utils": "6.6.0",
     "acorn": "8.10.0",
@@ -138,6 +138,8 @@
     "eslint": "8.48.0",
     "eslint-import-resolver-typescript": "3.6.0",
     "eslint-plugin-import": "npm:@phryneas/eslint-plugin-import@2.27.5-pr.2813.2817.199971c",
+    "eslint-plugin-local-rules": "2.0.0",
+    "eslint-plugin-testing-library": "5.11.1",
     "expect-type": "0.16.0",
     "fetch-mock": "9.11.0",
     "glob": "8.1.0",
