+++ tsconfig.tests.json
@@ -1,0 +1,5 @@
+{
+  "extends": "./tsconfig.json",
+  "include": ["src/**/__tests__/**/*.ts", "src/**/__tests__/**/*.tsx"],
+  "exclude": []
+}
